feat: config reorganization
