+++ assets/logo.sh
