+++ lua/plugins/ui.lua
@@ -1,10 +1,10 @@
 return {
     {
         "nvim-lualine/lualine.nvim",
-        dependencies = {},
+        event = "VeryLazy",
         config = function()
             local noice = require("noice")
-            local colors = require("tokyonight.colors").setup({})
+            local colors = require("tokyonight.colors").setup()
             local theme = require("lualine.themes.tokyonight")
 
             -- Change the background of lualine_c section for normal mode
@@ -35,7 +35,7 @@ return {
                             separator = "%#StatusLineSeparator#·%#StatusLine#",
                         },
                         {
-                            require("plugins.tools.helpers").lsp_breakdown,
+                            require("helpers.lualine_lsp").map_lsp_to_info,
                             separator = "    ",
                         },
                         {
@@ -135,19 +135,17 @@ return {
 
     {
         "yamatsum/nvim-cursorline",
-        event = "VeryLazy",
-        config = function()
-            require("nvim-cursorline").setup({
-                cursorline = {
-                    enable = true,
-                    timeout = 0,
-                    number = false,
-                },
-                cursorword = {
-                    enable = false,
-                },
-            })
-        end,
+        event = { "BufReadPost", "BufNewFile" },
+        opts = {
+            cursorline = {
+                enable = true,
+                timeout = 0,
+                number = false,
+            },
+            cursorword = {
+                enable = false,
+            },
+        },
     },
 
     {
@@ -157,131 +155,143 @@ return {
             "rcarriga/nvim-notify",
         },
         event = "VeryLazy",
-        config = function()
-            require("notify").setup({
-                stages = "static",
-            })
-
+        keys = function()
             local noice = require("noice")
             local noice_lsp = require("noice.lsp")
 
-            noice.setup({
-                lsp = {
-                    hover = {
-                        enabled = false,
-                    },
-                    signature = {
-                        enabled = false,
-                    },
-                    override = {
-                        ["vim.lsp.util.convert_input_to_markdown_lines"] = true,
-                        ["vim.lsp.util.stylize_markdown"] = true,
-                        ["cmp.entry.get_documentation"] = true,
-                    },
+            return {
+                {
+                    "<leader>nl",
+                    function()
+                        noice.cmd("last")
+                    end,
+                    desc = "Last message in popup",
                 },
-                presets = {
-                    bottom_search = true,
-                    command_palette = true,
-                    long_message_to_split = true,
-                    inc_rename = false,
-                    lsp_doc_border = false,
+                {
+                    "<leader>nh",
+                    function()
+                        noice.cmd("history")
+                    end,
+                    desc = "Message history",
                 },
-                views = {
-                    mini = {
-                        win_options = {
-                            winblend = 0
-                        }
-                    }
-                }
-            })
-
-            require("telescope").load_extension("noice")
-
-            vim.keymap.set("n", "<leader>nl", function()
-                noice.cmd("last")
-            end, { desc = "Last message in popup" })
-
-            vim.keymap.set("n", "<leader>nh", function()
-                noice.cmd("history")
-            end, { desc = "Message history" })
-
-            vim.keymap.set("c", "<S-Enter>", function()
-                noice.redirect(vim.fn.getcmdline())
-            end, { desc = "Redirect Cmdline" })
-
-            vim.keymap.set({ "n", "i", "s" }, "<c-f>", function()
-                if not noice_lsp.scroll(4) then
-                    return "<c-f>"
-                end
-            end, {
-                silent = true,
-                expr = true,
-                desc = "Scroll history forward (LSP)",
-            })
-            vim.keymap.set(
-                "n",
-                "<leader>nt",
-                "<cmd>Telescope noice<cr>",
-                { desc = "Telescope Noice" }
-            )
-
-            vim.keymap.set({ "n", "i", "s" }, "<c-b>", function()
-                if not noice_lsp.scroll(-4) then
-                    return "<c-b>"
-                end
-            end, {
-                silent = true,
-                expr = true,
-                desc = "Scroll history backward (LSP)",
-            })
+                {
+                    "<S-Enter>",
+                    function()
+                        noice.redirect(vim.fn.getcmdline())
+                    end,
+                    mode = { "c" },
+                    desc = "Redirect Cmdline",
+                },
+                {
+                    "<c-f>",
+                    function()
+                        if not noice_lsp.scroll(4) then
+                            return "<c-f>"
+                        end
+                    end,
+                    mode = { "n", "i", "s" },
+                    desc = "Scroll history forward (LSP)",
+                    silent = true,
+                    expr = true,
+                },
+                {
+                    "<leader>nt",
+                    "<cmd>Telescope noice<cr>",
+                    desc = "Telescope Noice",
+                },
+                {
+                    "<c-b>",
+                    function()
+                        if not noice_lsp.scroll(-4) then
+                            return "<c-b>"
+                        end
+                    end,
+                    mode = { "n", "i", "s" },
+                    desc = "Scroll history backward (LSP)",
+                    silent = true,
+                    expr = true,
+                },
+            }
         end,
+        opts = {
+            lsp = {
+                hover = {
+                    enabled = false,
+                },
+                signature = {
+                    enabled = false,
+                },
+                override = {
+                    ["vim.lsp.util.convert_input_to_markdown_lines"] = true,
+                    ["vim.lsp.util.stylize_markdown"] = true,
+                    ["cmp.entry.get_documentation"] = true,
+                },
+            },
+            presets = {
+                bottom_search = true,
+                command_palette = true,
+                long_message_to_split = true,
+                inc_rename = false,
+                lsp_doc_border = false,
+            },
+            views = {
+                mini = {
+                    win_options = {
+                        winblend = 0,
+                    },
+                },
+            },
+        },
+    },
+
+    {
+        "rcarriga/nvim-notify",
+        lazy = true,
+        opts = {
+            stages = "static",
+        },
     },
 
     {
         "nmac427/guess-indent.nvim",
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require("guess-indent").setup({})
-        end,
+        event = { "BufReadPost", "BufNewFile" },
+        config = true,
     },
 
     {
         "lukas-reineke/indent-blankline.nvim",
-        event = { "BufReadPre", "BufNewFile" },
+        event = { "BufReadPost", "BufNewFile" },
         dependencies = {
             "https://gitlab.com/HiPhish/rainbow-delimiters.nvim.git",
         },
-        config = function()
+        init = function()
             vim.opt.list = true
-
-            local indent_highlights = {
-                "RainbowLight1",
-                "RainbowLight2",
-                "RainbowLight3",
-                "RainbowLight4",
-                "RainbowLight5",
-                "RainbowLight6",
-                "RainbowLight7",
-            }
-
-            local scope_highlights = {
-                "Rainbow1",
-                "Rainbow2",
-                "Rainbow3",
-                "Rainbow4",
-                "Rainbow5",
-                "Rainbow6",
-                "Rainbow7",
-            }
-
+        end,
+        config = function()
             require("ibl").setup({
                 indent = {
-                    highlight = indent_highlights,
+                    highlight = {
+                        "RainbowLight1",
+                        "RainbowLight2",
+                        "RainbowLight3",
+                        "RainbowLight4",
+                        "RainbowLight5",
+                        "RainbowLight6",
+                        "RainbowLight7",
+                    },
                 },
                 scope = {
                     show_start = false,
                     show_end = false,
-                    highlight = scope_highlights,
+                    highlight = {
+                        "Rainbow1",
+                        "Rainbow2",
+                        "Rainbow3",
+                        "Rainbow4",
+                        "Rainbow5",
+                        "Rainbow6",
+                        "Rainbow7",
+                    },
                 },
             })
         end,
@@ -291,19 +301,17 @@ return {
         "https://gitlab.com/HiPhish/rainbow-delimiters.nvim.git",
         lazy = true,
         config = function()
-            local scope_highlights = {
-                "Rainbow1",
-                "Rainbow2",
-                "Rainbow3",
-                "Rainbow4",
-                "Rainbow5",
-                "Rainbow6",
-                "Rainbow7",
-            }
-
             require("rainbow-delimiters")
             vim.g.rainbow_delimiters = {
-                highlight = scope_highlights,
+                highlight = {
+                    "Rainbow1",
+                    "Rainbow2",
+                    "Rainbow3",
+                    "Rainbow4",
+                    "Rainbow5",
+                    "Rainbow6",
+                    "Rainbow7",
+                },
             }
         end,
     },
@@ -311,43 +319,43 @@ return {
     {
         "folke/todo-comments.nvim",
         event = { "BufReadPost", "BufNewFile" },
-        config = function()
-            require("todo-comments").setup({
-                sign_priority = 10,
-            })
-
-            vim.keymap.set("n", "]t", function()
-                require("todo-comments").jump_next()
-            end, { desc = "Next todo comment" })
-
-            vim.keymap.set("n", "[t", function()
-                require("todo-comments").jump_prev()
-            end, { desc = "Previous todo comment" })
-        end,
+        keys = {
+            {
+                "]t",
+                function()
+                    require("todo-comments").jump_next()
+                end,
+                desc = "Next todo comment",
+            },
+            {
+                "[t",
+                function()
+                    require("todo-comments").jump_prev()
+                end,
+                desc = "Previous todo comment",
+            },
+        },
+        opts = {
+            sign_priority = 10,
+        },
     },
 
     {
         "tzachar/highlight-undo.nvim",
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require("highlight-undo").setup({
-                duration = 150,
-            })
-        end,
+        event = { "BufReadPost", "BufNewFile" },
+        opts = {
+            duration = 150,
+        },
     },
 
     {
         "NvChad/nvim-colorizer.lua",
-        event = { "BufReadPre", "BufNewFile" },
-        config = function()
-            require("colorizer").setup({})
-        end,
+        event = { "BufReadPost", "BufNewFile" },
+        config = true,
     },
 
     {
         "Bekaboo/deadcolumn.nvim",
-        event = { "BufReadPre", "BufNewFile" },
+        event = { "BufReadPost", "BufNewFile" },
     },
-
-    { "nvim-tree/nvim-web-devicons", lazy = true },
 }
