+++ lua/set.lua
@@ -1,6 +1,8 @@
 vim.opt.autoread = true
 vim.cmd([[autocmd CursorHold * checktime]])
-vim.cmd([[autocmd TextYankPost * silent! lua vim.highlight.on_yank({higroup="Yanked"})]])
+vim.cmd(
+    [[autocmd TextYankPost * silent! lua vim.highlight.on_yank({higroup="Yanked"})]]
+)
 
 vim.opt.nu = true
 vim.opt.relativenumber = true
